+++ source/renderer/app/components/assets/Asset.js
@@ -173,7 +173,7 @@ export default class Asset extends Component<Props, State> {
     const {
       asset,
       metadataNameChars,
-      small,
+      small = true,
       fullFingerprint,
       hasWarning,
       hasError,
@@ -182,9 +182,9 @@ export default class Asset extends Component<Props, State> {
     const { name } = metadata || {};
     const contentStyles = classnames([
       styles.pill,
-      small ? styles.small : null,
       hasError ? styles.error : null,
     ]);
+    const [startCharAmount, endCharAmoout] = small ? [9, 4] : [12, 12];
     let warningPopOverMessage;
     if (hasWarning) {
       warningPopOverMessage =
@@ -195,7 +195,9 @@ export default class Asset extends Component<Props, State> {
     return (
       <div className={contentStyles}>
         <div className={styles.fingerprint}>
-          {fullFingerprint ? fingerprint : ellipsis(fingerprint || '', 9, 4)}
+          {fullFingerprint
+            ? fingerprint
+            : ellipsis(fingerprint || '', startCharAmount, endCharAmoout)}
         </div>
         {name && (
           <div className={styles.metadataName}>
@@ -223,7 +225,6 @@ export default class Asset extends Component<Props, State> {
       asset,
       onCopyAssetParam,
       assetSettingsDialogWasOpened,
-      anyAssetWasHovered,
     } = this.props;
     const pillContent = this.renderPillContent();
     const popOverContent = (
@@ -231,7 +232,6 @@ export default class Asset extends Component<Props, State> {
         asset={asset}
         onCopyAssetParam={onCopyAssetParam}
         assetSettingsDialogWasOpened={assetSettingsDialogWasOpened}
-        anyAssetWasHovered={anyAssetWasHovered}
         className={styles.popOverContent}
         highlightFingerprint
       />